Move `PublishResponse` command
